% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.R
\name{cube}
\alias{cube}
\title{cube}
\usage{
cube(size, center = FALSE)
}
\arguments{
\item{size}{the size of the cube. Use a single value for a cube or a vector of three values (x,y,z) for a cuboid}

\item{center}{center the cube on the x,y,z origin}
}
\description{
cube
}
