% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.R
\name{cylinder}
\alias{cylinder}
\title{create a cylinder}
\usage{
cylinder(height, radius, radius2 = NA, center = FALSE)
}
\arguments{
\item{height}{height}

\item{radius}{radius}

\item{radius2}{secondary radius for cone-like 3D shape}

\item{center}{bool}
}
\description{
create a cylinder
}
