% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d.R
\name{linear_extrude}
\alias{linear_extrude}
\title{Linear extrusion operations}
\usage{
linear_extrude(
  object,
  height,
  center = FALSE,
  convexity = 10,
  twist = 0,
  slices = NA,
  scale = 1
)
}
\arguments{
\item{object}{a 2D openSCAD object}

\item{height}{the height of the extrusion}

\item{center}{Center on the origin}

\item{convexity}{convexity level}

\item{twist}{degrees of twist}

\item{slices}{number of intermediate points along the Z axis of the extrusion}

\item{scale}{scale}
}
\description{
Takes a 2 dimensional object as input and generates a 3d object as output
}
