% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_to_stl}
\alias{model_to_stl}
\title{Convert a model file to an STL file}
\usage{
model_to_stl(model, outfile)
}
\arguments{
\item{model}{A model to convert to STL}

\item{outfile}{the STL file to write out}
}
\description{
Convert a model file to an STL file
}
