% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_in_OpenSCAD}
\alias{open_in_OpenSCAD}
\title{Open the given file in the OpenSCAD editor}
\usage{
open_in_OpenSCAD(filename)
}
\arguments{
\item{filename}{the name of the SCAD file to open in the OpenSCAD editor}
}
\description{
Open the given file in the OpenSCAD editor
}
