% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preview}
\alias{preview}
\title{Generate a preview of the given SCAD file}
\usage{
preview(filename)
}
\arguments{
\item{filename}{filename from which to generate a preview}
}
\description{
Generate a preview of the given SCAD file
}
