% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{resize}
\alias{resize}
\title{resize a 3D object}
\usage{
resize(object, x = 0, y = 0, z = 0, auto = FALSE, convexity = 10)
}
\arguments{
\item{object}{object to resize}

\item{x}{amount to resize on the x axis}

\item{y}{amount to resize on the y axis}

\item{z}{amount to resize on the z axis}

\item{auto}{If the 'auto' parameter is set to true, it auto-scales any 0-dimensions to match}

\item{convexity}{convexity level}
}
\description{
resize a 3D object
}
