% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rotate}
\alias{rotate}
\title{Rotate on object in 3D space}
\usage{
rotate(object, x = 0, y = 0, z = 0)
}
\arguments{
\item{object}{an openscad object to translate}

\item{x}{amount to rotate the object by on the x axis}

\item{y}{amount to rotate the object by on the y axis}

\item{z}{amount to rotate the object by on the z axis}
}
\description{
Rotate on object in 3D space
}
