% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scad_to_stl}
\alias{scad_to_stl}
\title{Convert a SCAD file to an STL file}
\usage{
scad_to_stl(scadfile, outfile)
}
\arguments{
\item{scadfile}{A scad file to convert to STL}

\item{outfile}{the STL file to write out}
}
\description{
Convert a SCAD file to an STL file
}
