% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{topo_matrix}
\alias{topo_matrix}
\title{Convert a topographical matrix to SCAD}
\usage{
topo_matrix(
  topology,
  block_size = 5,
  scaling = 1,
  flatten = FALSE,
  base_thickness = 0
)
}
\arguments{
\item{topology}{a matrix containing topological values (like R's built-in volcano)}

\item{block_size}{The size of the squares in the SCAD file}

\item{scaling}{a scaling factor to increase of decrease the vertical scale of the blocks}

\item{flatten}{Where all defined values in the topology are above x, subtract x from those values}

\item{base_thickness}{Adds additional depth to the base}
}
\description{
Convert a topographical matrix to SCAD
}
