% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{translate}
\alias{translate}
\title{translate an objects position in 3D space}
\usage{
translate(object, x = 0, y = 0, z = 0)
}
\arguments{
\item{object}{an openscad object to translate}

\item{x}{amount to move the object by on the x axis}

\item{y}{amount to move the object by on the y axis}

\item{z}{amount to move the object by on the z axis}
}
\description{
translate an objects position in 3D space
}
