% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_scad}
\alias{write_scad}
\title{Write scad object out to file}
\usage{
write_scad(object, filename, overwrite = FALSE)
}
\arguments{
\item{object}{an openscad object to write to the filesystem}

\item{filename}{the filename to write}

\item{overwrite}{Controls whether to overwrite an existing file}
}
\description{
Write scad object out to file
}
